platina-goes-release: Remove submodules and switch to shallow clones
Replace the use of submodules with a script to check out the git
repositories using the most efficient mode possible.

Alas, Platina components currently use information from git for
versioning in the build process, which restricts our use of shallow
clones.

Fixing this is TBD. This is a first pass.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>